Add files via upload
The given sample dataset is related to a housing dataset given by a Real Estate
Agency with various features that could influence the housing prices. The
dataset contains various independent attributes which impact the dependent
variable. The data set contains Crime rate, Age, Distance, Tax, Avg room, Avg
Price. The brief summary of the above mentioned variables is given below:
a. Crime Rate: This column represents the crime rate in different locations.
b. Age: This represents the age in the locality.
c. Distance: This represents distance between the house locality and center of
the city.
d. Tax: It represents property tax rate.
e. Avg Room: This column represents number of rooms per house.
The above detailed variables are the independent variables which does have the
impact on the below mentioned dependent variable.
f. Avg Price: It represents the average prices of houses in the locality.